Rename: Bithumb 컴포넌트 폴더 이동
